refactor: List, Detail 스타일 코드 중복 제거 및 모듈화
