+++ src/App.tsx
@@ -1,7 +1,7 @@
 import DefaultLayout from 'layout/defaultLayout';
 import Bus from 'pages/Services/Bus';
 import Cafeteria from 'pages/Services/Cafeteria';
-import Category from 'pages/Services/Category';
+import Category from 'pages/Services/Category/CategoryList';
 import Dashboard from 'pages/Dashboard';
 import Login from 'pages/Login';
 import Manager from 'pages/UserManage/Manager';
