+++ src/pages/Services/Category/CategoryDetail.tsx
@@ -1,0 +1,10 @@
+import CustomForm from 'components/common/CustomForm';
+
+export default function CategoryDetail() {
+  return (
+    <div>
+      <h1>Category Detail</h1>
+      <CustomForm />
+    </div>
+  );
+}
