+++ src/pages/Services/Category/CategoryList.style.tsx
@@ -1,0 +1,1 @@
+export * from 'styles/List.style';
