+++ src/pages/UserManage/Member/MemberDetail.style.tsx
@@ -1,18 +1,6 @@
 import styled from 'styled-components';
 
-export const Container = styled.div`
-  font-family: 'Avenir', 'Helvetica', 'Arial', 'sans-serif';
-`;
-
-export const Heading = styled.div`
-  color: #2c3e50;
-  font-size: 28px;
-`;
-
-export const BreadCrumb = styled.div`
-  font-size: 13px;
-  padding: 12px 0;
-`;
+export * from 'styles/Detail.style';
 
 export const FormWrapper = styled.div`
   padding: 10px 120px 60px 20px;
@@ -21,14 +9,4 @@ export const FormWrapper = styled.div`
     font-weight: 600;
     color: #2c3e50;
   }
-`;
-
-export const ButtonWrap = styled.div`
-  margin-top: 20px;
-  margin-left: 15px;
-  display: flex;
-  justify-content: flex-end;
-`;
-
-export const UploadWrap = styled.div`
 `;
