+++ src/pages/UserManage/Member/MemberList.style.tsx
@@ -1,11 +1,6 @@
 import styled from 'styled-components';
 
-export const Heading = styled.h1`
-  font-size: 30px;
-  font-weight: 700;
-  color: #404040;
-  padding: 12px 0 0 12px;
-`;
+export { Heading } from 'styles/List.style';
 
 export const Tab = styled.button<{ selected: boolean }>`
   appearance: none;
