+++ src/pages/UserManage/User/UserList.tsx
@@ -1,7 +1,7 @@
 import CustomTable from 'components/common/CustomTable';
 import { useState } from 'react';
 import { useGetUserListQuery } from 'store/api/user';
-import * as S from 'styles/List.style';
+import * as S from './UserList.style';
 
 function UserList() {
   const [page, setPage] = useState(1);
